Move the unit tests
