+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/PaymentSheetAddPaymentMethodFragmentTest.kt
@@ -315,58 +315,6 @@ internal class PaymentSheetAddPaymentMethodFragmentTest : PaymentSheetViewModelT
     }
 
     @Test
-    fun `card payment method selection has the fields from formFieldValues`() {
-        val formFieldValues = FormFieldValues(
-            fieldValuePairs = mapOf(
-                IdentifierSpec.SaveForFutureUse to FormFieldEntry("true", true),
-                IdentifierSpec.CardNumber to FormFieldEntry("4242424242421234", true),
-                IdentifierSpec.CardBrand to FormFieldEntry(CardBrand.Visa.code, true)
-            ),
-            showsMandate = false,
-            userRequestedReuse = PaymentSelection.CustomerRequestedSave.RequestReuse
-        )
-        val selection =
-            BaseAddPaymentMethodFragment.transformToPaymentSelection(
-                formFieldValues,
-                SupportedPaymentMethod.Card
-            )
-        assertThat(selection?.customerRequestedSave).isEqualTo(
-            PaymentSelection.CustomerRequestedSave.RequestReuse
-        )
-        assertThat((selection as? PaymentSelection.New.Card)?.last4).isEqualTo(
-            "1234"
-        )
-        assertThat((selection as? PaymentSelection.New.Card)?.brand).isEqualTo(
-            CardBrand.Visa
-        )
-    }
-
-    @Test
-    fun `payment method selection has the fields from formFieldValues`() {
-        val formFieldValues = FormFieldValues(
-            fieldValuePairs = mapOf(
-                IdentifierSpec.SaveForFutureUse to FormFieldEntry("true", true)
-            ),
-            showsMandate = false,
-            userRequestedReuse = PaymentSelection.CustomerRequestedSave.RequestReuse
-        )
-        val selection =
-            BaseAddPaymentMethodFragment.transformToPaymentSelection(
-                formFieldValues,
-                SupportedPaymentMethod.Sofort
-            )
-        assertThat(selection?.customerRequestedSave).isEqualTo(
-            PaymentSelection.CustomerRequestedSave.RequestReuse
-        )
-        assertThat((selection as? PaymentSelection.New.GenericPaymentMethod)?.labelResource).isEqualTo(
-            R.string.stripe_paymentsheet_payment_method_sofort
-        )
-        assertThat((selection as? PaymentSelection.New.GenericPaymentMethod)?.iconResource).isEqualTo(
-            R.drawable.stripe_ic_paymentsheet_pm_klarna
-        )
-    }
-
-    @Test
     fun `Factory gets initialized by Injector when Injector is available`() {
         createFragment(registerInjector = true) { fragment, _, viewModel ->
             assertThat(fragment.sheetViewModel).isEqualTo(viewModel)
